Refactor application page models a bit
