+++ frontend/src/e2e-playwright/pages/employee/applications.ts
@@ -2,8 +2,9 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import { Page, Checkbox, FileInput } from 'e2e-playwright/utils/page'
-import { waitUntilEqual, waitUntilFalse } from 'e2e-playwright/utils'
+import { Element, Page } from 'e2e-playwright/utils/page'
+import ApplicationReadView from './applications/application-read-view'
+import { waitUntilEqual } from '../../utils'
 
 export default class ApplicationsPage {
   constructor(private readonly page: Page) {}
@@ -18,18 +19,7 @@ export default class ApplicationsPage {
 
   applicationRow(id: string) {
     const element = this.page.find(`[data-application-id="${id}"]`)
-    return {
-      status: element.find('[data-qa="application-status"]'),
-      openApplication: async () => {
-        const applicationDetails = new Promise<ApplicationDetailsPage>(
-          (res) => {
-            this.page.onPopup((page) => res(new ApplicationDetailsPage(page)))
-          }
-        )
-        await element.click()
-        return applicationDetails
-      }
-    }
+    return new ApplicationRow(this.page, element)
   }
 
   readonly details = {
@@ -37,98 +27,22 @@ export default class ApplicationsPage {
   }
 }
 
-class ApplicationDetailsPage {
-  constructor(private readonly page: Page) {}
-
-  #editButton = this.page.find('[data-qa="edit-application"]')
-  #saveButton = this.page.find('[data-qa="save-application"]')
-  #urgentCheckbox = new Checkbox(this.page.find('[data-qa="checkbox-urgent"]'))
-  #urgentAttachmentFileUpload = this.page.find('[data-qa="file-upload-urgent"]')
-  #shiftCareCheckbox = new Checkbox(
-    this.page.find('[data-qa="checkbox-service-need-shift-care"]')
-  )
-  #shiftCareAttachmentFileUpload = this.page.find(
-    '[data-qa="file-upload-shift-care"]'
-  )
-
-  applicantDeadIndicator = this.page.find('[data-qa="applicant-dead"]')
-
-  async startEditing() {
-    await this.#editButton.click()
+export class ApplicationRow extends Element {
+  constructor(private page: Page, root: Element) {
+    super(root)
   }
 
-  async saveApplication() {
-    await this.#saveButton.click()
+  #status = this.find('[data-qa="application-status"]')
+
+  async openApplication() {
+    const applicationDetails = new Promise<ApplicationReadView>((res) => {
+      this.page.onPopup((page) => res(new ApplicationReadView(page)))
+    })
+    await this.click()
+    return applicationDetails
   }
 
-  async setUrgent() {
-    if (await this.#urgentCheckbox.checked) {
-      return
-    }
-    await this.#urgentCheckbox.click()
-  }
-
-  async uploadUrgentAttachment(filePath: string) {
-    await new FileInput(
-      this.#urgentAttachmentFileUpload.find('[data-qa="btn-upload-file"]')
-    ).setInputFiles(filePath)
-  }
-
-  async assertUrgentAttachmentUploaded(fileName: string) {
-    await waitUntilEqual(
-      () =>
-        this.#urgentAttachmentFileUpload.find(
-          '[data-qa="file-download-button"]'
-        ).innerText,
-      fileName
-    )
-  }
-
-  async assertUrgencyAttachmentReceivedAtVisible(fileName: string) {
-    const attachment = this.page.find(
-      `[data-qa="urgent-attachment-${fileName}"]`
-    )
-    await attachment.waitUntilVisible()
-    await attachment
-      .find('[data-qa="attachment-received-at"]')
-      .waitUntilVisible()
-  }
-
-  async setShiftCareNeeded() {
-    if (await this.#shiftCareCheckbox.checked) {
-      return
-    }
-    await this.#shiftCareCheckbox.click()
-  }
-
-  async uploadShiftCareAttachment(filePath: string) {
-    await new FileInput(
-      this.#shiftCareAttachmentFileUpload.find('[data-qa="btn-upload-file"]')
-    ).setInputFiles(filePath)
-  }
-
-  async deleteShiftCareAttachment(fileName: string) {
-    await this.#shiftCareAttachmentFileUpload
-      .find(`[data-qa="file-delete-button-${fileName}"]`)
-      .click()
-  }
-
-  async assertShiftCareAttachmentUploaded(fileName: string) {
-    await waitUntilEqual(
-      () =>
-        this.#shiftCareAttachmentFileUpload.find(
-          '[data-qa="file-download-button"]'
-        ).innerText,
-      fileName
-    )
-  }
-
-  async assertShiftCareAttachmentsDeleted() {
-    await waitUntilFalse(
-      () =>
-        this.#shiftCareAttachmentFileUpload.find(
-          '[data-qa="file-download-button"]'
-        ).visible
-    )
+  async assertStatus(expectedStatus: string) {
+    await waitUntilEqual(() => this.#status.innerText, expectedStatus)
   }
 }
